add logger

+++ AStream-go/utils/file.go
@@ -10,6 +10,9 @@ import (
 
 func CleanFiles(dirPath string) {
 	err := filepath.Walk(dirPath, func(filePath string, info fs.FileInfo, err error) error {
+		if info.IsDir() {
+			return nil
+		}
 		return os.Remove(filePath)
 	})
 	if err != nil {
